requirements.txt
# ── Core ─────────────────────────────────────────────────────
pandas==2.2.2
numpy==1.26.4
pyyaml==6.0.1
python-dotenv==1.0.1

# ── NLP ──────────────────────────────────────────────────────
scikit-learn==1.4.2
spacy==3.7.4
bertopic==0.16.2
sentence-transformers==2.7.0

# ── Database ─────────────────────────────────────────────────
sqlalchemy==2.0.30
psycopg2-binary==2.9.9

# ── Dashboard ────────────────────────────────────────────────
dash==2.17.1
plotly==5.22.0
dash-bootstrap-components==1.6.0

# ── LLM ──────────────────────────────────────────────────────
anthropic>=0.40.0
groq>=0.9.0

# ── Testing ──────────────────────────────────────────────────
pytest==8.2.2
